🌱 refactor: split WhatsNewModal into sub-components (partial #21171)
Extract four focused units from the 436-line WhatsNewModal:
- useRecentPRs: PR-fetching hook (useEffect + 3 state vars)
- ReleaseNotesContent: notes/PR-list/loading/error display
- PreviousReleases: collapsible section, owns its own open/expand state
- WhatsNewFooter: skip/snooze/update buttons, owns snooze-menu state

WhatsNewModal drops to 216 lines and becomes a pure composition root.
No functional or UI changes — behavior is preserved 1:1.
