+++ web/src/components/updates/WhatsNewModal.tsx
@@ -1,13 +1,8 @@
 import { COPY_FEEDBACK_TIMEOUT_MS } from '../../lib/constants'
 import { formatTimeAgo } from '../../lib/formatters'
-import { useState, useMemo, useCallback, useEffect } from 'react'
-import { Download, Clock, SkipForward, ChevronDown, Copy, Check, Loader2 } from 'lucide-react'
-import { LazyMarkdown as ReactMarkdown } from '../ui/LazyMarkdown'
-import remarkGfm from 'remark-gfm'
-import remarkBreaks from 'remark-breaks'
-import rehypeSanitize from 'rehype-sanitize'
+import { useState, useMemo, useCallback } from 'react'
+import { Download, ChevronDown, Copy, Check } from 'lucide-react'
 import { BaseModal } from '../../lib/modals'
-import { MS_PER_HOUR, MS_PER_DAY } from '../../lib/constants/time'
 import { useVersionCheck } from '../../hooks/useVersionCheck'
 import { useSelfUpgrade } from '../../hooks/useSelfUpgrade'
 import { useToast } from '../ui/Toast'
@@ -15,19 +10,17 @@ import { useTranslation } from 'react-i18next'
 import { buildReleaseNotesComponents } from '../../lib/markdown/releaseNotesComponents'
 import { cn } from '../../lib/cn'
 import { copyToClipboard } from '../../lib/clipboard'
-import { FETCH_DEFAULT_TIMEOUT_MS } from '../../lib/constants'
-import { api, RateLimitError } from '../../lib/api'
 import { emitWhatsNewUpdateClicked, emitWhatsNewRemindLater } from '../../lib/analytics'
+import { useRecentPRs } from './whatsNew/useRecentPRs'
+import { ReleaseNotesContent } from './whatsNew/ReleaseNotesContent'
+import { PreviousReleases } from './whatsNew/PreviousReleases'
+import { WhatsNewFooter } from './whatsNew/WhatsNewFooter'
 
 const SNOOZE_STORAGE_KEY = 'kc-update-snoozed'
 
 // Clean up the removed kill-switch key so users who had it set during
 // development don't get a dead button. Runs once on module load.
 try { localStorage.removeItem('kc-whats-new-modal-disabled') } catch { /* ignore */ }
-
-const SNOOZE_DURATION_1H_MS = MS_PER_HOUR
-const SNOOZE_DURATION_1D_MS = MS_PER_DAY
-const SNOOZE_DURATION_1W_MS = 7 * MS_PER_DAY
 
 export function isUpdateSnoozed(): boolean {
   try {
@@ -40,7 +33,6 @@ export function isUpdateSnoozed(): boolean {
   }
 }
 
-
 function snoozeUpdate(durationMs: number) {
   try {
     localStorage.setItem(SNOOZE_STORAGE_KEY, String(Date.now() + durationMs))
@@ -48,8 +40,6 @@ function snoozeUpdate(durationMs: number) {
     // localStorage unavailable — silently ignore
   }
 }
-
-
 
 interface WhatsNewModalProps {
   isOpen: boolean
@@ -71,79 +61,13 @@ export function WhatsNewModal({ isOpen, onClose }: WhatsNewModalProps) {
   const { triggerUpgrade } = useSelfUpgrade()
 
   const [updating, setUpdating] = useState(false)
-  const [showPreviousReleases, setShowPreviousReleases] = useState(false)
-  const [expandedRelease, setExpandedRelease] = useState<string | null>(null)
   const [showManualUpdate, setShowManualUpdate] = useState(false)
-  const [showSnoozeMenu, setShowSnoozeMenu] = useState(false)
   const [copiedCommand, setCopiedCommand] = useState<string | null>(null)
-  const [recentPRs, setRecentPRs] = useState<Array<{ number: number; title: string; merged_at: string }>>([])
-  // Auto-QA #9036 — surface loading/error state for the PR fallback fetch
-  // so users see either a spinner while the list is being fetched or a
-  // short error message if GitHub is unreachable, instead of an empty modal.
-  const [prsLoading, setPrsLoading] = useState(false)
-  const [prsError, setPrsError] = useState<string | null>(null)
 
   const markdownComponents = useMemo(() => buildReleaseNotesComponents('sm'), [])
-
-  // When release notes are empty, fetch merged PRs since the running commit.
-  // First get the commit date for the running hash, then filter PRs merged after.
-  const MAX_RECENT_PRS = 50
   const hasReleaseNotes = !!latestRelease?.releaseNotes?.trim()
-  useEffect(() => {
-    if (hasReleaseNotes || !isOpen) return
-    let cancelled = false
-    const fetchPRs = async () => {
-      setPrsLoading(true)
-      setPrsError(null)
-      try {
-        // Step 1: Get the date of the currently running commit
-        let sinceDate: string | null = null
-        if (commitHash && commitHash !== 'unknown') {
-          try {
-            const { data: commitData } = await api.get<{ commit?: { committer?: { date?: string } } }>(
-              `/api/github/repos/kubestellar/console/commits/${commitHash}`,
-              { timeout: FETCH_DEFAULT_TIMEOUT_MS },
-            )
-            sinceDate = commitData?.commit?.committer?.date ?? null
-          } catch {
-            // Commit lookup failure is non-fatal — show all PRs
-          }
-        }
 
-        // Step 2: Fetch recent merged PRs
-        if (cancelled) return
-        const { data } = await api.get<Array<{ number: number; title: string; merged_at: string | null }>>(
-          `/api/github/repos/kubestellar/console/pulls?state=closed&sort=updated&direction=desc&per_page=${MAX_RECENT_PRS}`,
-          { timeout: FETCH_DEFAULT_TIMEOUT_MS },
-        )
-        if (cancelled) return
-        const merged = (data || [])
-          .filter((pr) => pr.merged_at)
-          .filter((pr) => {
-            if (!sinceDate) return true
-            return new Date(pr.merged_at!) > new Date(sinceDate)
-          })
-          .map((pr) => ({
-            number: pr.number,
-            title: pr.title,
-            merged_at: pr.merged_at!,
-          }))
-        setRecentPRs(merged)
-      } catch (err: unknown) {
-        if (cancelled) return
-        if (err instanceof RateLimitError) {
-          setPrsError('GitHub API rate limit reached — data will refresh automatically. Try again in a moment.')
-          return
-        }
-        const message = err instanceof Error ? err.message : 'Network error'
-        setPrsError(message)
-      } finally {
-        if (!cancelled) setPrsLoading(false)
-      }
-    }
-    fetchPRs()
-    return () => { cancelled = true }
-  }, [hasReleaseNotes, isOpen, commitHash])
+  const { recentPRs, prsLoading, prsError } = useRecentPRs({ hasReleaseNotes, isOpen, commitHash })
 
   const previousReleases = useMemo(() => {
     if (!releases || !currentVersion || !latestRelease) return []
@@ -189,7 +113,6 @@ export function WhatsNewModal({ isOpen, onClose }: WhatsNewModalProps) {
 
   const handleSnooze = useCallback((durationMs: number, label: string) => {
     snoozeUpdate(durationMs)
-    setShowSnoozeMenu(false)
     emitWhatsNewRemindLater(latestRelease?.tag ?? '', label)
     showToast('Update reminder snoozed', 'info')
     onClose()
@@ -235,92 +158,17 @@ export function WhatsNewModal({ isOpen, onClose }: WhatsNewModalProps) {
       <BaseModal.Content>
         <div className="space-y-4">
           {/* Primary release notes — or recent merged PRs as fallback */}
-          {hasReleaseNotes ? (
-            <div className="prose dark:prose-invert max-w-none text-sm overflow-x-auto wrap-break-word [word-break:break-word] prose-pre:my-5 prose-pre:bg-transparent prose-pre:p-0 prose-code:text-purple-700 dark:prose-code:text-purple-300 prose-code:bg-black/5 dark:prose-code:bg-black/20 prose-code:px-1 prose-code:rounded">
-              <ReactMarkdown
-                remarkPlugins={[remarkGfm, remarkBreaks]}
-                rehypePlugins={[rehypeSanitize]}
-                components={markdownComponents}
-              >
-                {latestRelease?.releaseNotes ?? ''}
-              </ReactMarkdown>
-            </div>
-          ) : prsLoading ? (
-            // Auto-QA #9036 — loading state for the PR fallback fetch
-            <div className="flex items-center justify-center gap-2 py-6 text-sm text-muted-foreground">
-              <Loader2 className="w-4 h-4 animate-spin" />
-              <span>{t('updates.loadingRecentPullRequests')}</span>
-            </div>
-          ) : prsError ? (
-            // Auto-QA #9036 — error state replaces the old silent failure
-            <div className="p-3 rounded-lg bg-red-500/10 border border-red-500/20 text-xs text-red-400">
-              Couldn&apos;t load recent pull requests: {prsError}
-            </div>
-          ) : recentPRs.length > 0 ? (
-            <div className="space-y-1">
-              <p className="text-xs text-muted-foreground mb-2">
-                Recently merged pull requests:
-              </p>
-              {recentPRs.map((pr) => (
-                <a
-                  key={pr.number}
-                  href={`https://github.com/kubestellar/console/pull/${pr.number}`}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  className="flex items-start gap-2 px-2 py-1.5 rounded hover:bg-secondary/50 transition-colors group"
-                >
-                  <span className="text-xs text-muted-foreground font-mono shrink-0">#{pr.number}</span>
-                  <span className="text-sm text-foreground group-hover:text-primary transition-colors">{pr.title}</span>
-                  <span className="text-xs text-muted-foreground ml-auto shrink-0">
-                    {formatTimeAgo(new Date(pr.merged_at))}
-                  </span>
-                </a>
-              ))}
-            </div>
-          ) : (
-            <div className="prose dark:prose-invert max-w-none text-sm">
-              <ReactMarkdown remarkPlugins={[remarkGfm, remarkBreaks]} rehypePlugins={[rehypeSanitize]} components={markdownComponents}>
-                {'*No release notes available. Pull the latest commits to update.*'}
-              </ReactMarkdown>
-            </div>
-          )}
+          <ReleaseNotesContent
+            hasReleaseNotes={hasReleaseNotes}
+            releaseNotes={latestRelease?.releaseNotes}
+            recentPRs={recentPRs}
+            prsLoading={prsLoading}
+            prsError={prsError}
+            markdownComponents={markdownComponents}
+          />
 
           {/* Previous releases (collapsible) */}
-          {previousReleases.length > 0 && (
-            <div className="border-t border-border pt-4">
-              <button
-                type="button"
-                onClick={() => setShowPreviousReleases(!showPreviousReleases)}
-                className="flex items-center gap-2 text-sm text-muted-foreground hover:text-foreground transition-colors w-full text-left"
-              >
-                <ChevronDown className={cn('w-4 h-4 transition-transform', showPreviousReleases && 'rotate-180')} />
-                Previous releases ({previousReleases.length} versions)
-              </button>
-              {showPreviousReleases && (
-                <div className="mt-3 space-y-2">
-                  {previousReleases.map(release => (
-                    <div key={release.tag} className="border border-border rounded-lg">
-                      <button
-                        type="button"
-                        onClick={() => setExpandedRelease(expandedRelease === release.tag ? null : release.tag)}
-                        className="flex items-center justify-between w-full px-3 py-2 text-left text-sm hover:bg-secondary/50 transition-colors rounded-lg"
-                      >
-                        <span className="font-medium text-foreground">{release.tag}</span>
-                        <span className="text-xs text-muted-foreground">{formatTimeAgo(release.publishedAt)}</span>
-                      </button>
-                      {expandedRelease === release.tag && release.releaseNotes && (
-                        <div className="px-3 pb-3 prose dark:prose-invert max-w-none text-xs overflow-x-auto wrap-break-word [word-break:break-word]">
-                          <ReactMarkdown remarkPlugins={[remarkGfm, remarkBreaks]} rehypePlugins={[rehypeSanitize]} components={markdownComponents}>
-                            {release.releaseNotes}
-                          </ReactMarkdown>
-                        </div>
-                      )}
-                    </div>
-                  ))}
-                </div>
-              )}
-            </div>
-          )}
+          <PreviousReleases releases={previousReleases} markdownComponents={markdownComponents} />
 
           {/* Manual update commands (collapsible) */}
           <div className="border-t border-border pt-4">
@@ -355,81 +203,13 @@ export function WhatsNewModal({ isOpen, onClose }: WhatsNewModalProps) {
       </BaseModal.Content>
 
       <BaseModal.Footer showKeyboardHints={false}>
-        <div className="flex items-center justify-between w-full">
-          <div className="flex items-center gap-2">
-            <button
-              type="button"
-              onClick={handleSkip}
-              className="px-3 py-1.5 text-xs text-muted-foreground hover:text-foreground transition-colors flex items-center gap-1"
-            >
-              <SkipForward className="w-3.5 h-3.5" />
-              Skip this version
-            </button>
-
-            {/* Remind me later dropdown */}
-            <div className="relative">
-              <button
-                type="button"
-                onClick={() => setShowSnoozeMenu(!showSnoozeMenu)}
-                className="px-3 py-1.5 text-xs text-muted-foreground hover:text-foreground transition-colors flex items-center gap-1"
-              >
-                <Clock className="w-3.5 h-3.5" />
-                Remind me later
-                <ChevronDown className="w-3 h-3" />
-              </button>
-              {showSnoozeMenu && (
-                <div className="absolute bottom-full left-0 mb-1 bg-card border border-border rounded-lg shadow-xl z-50 py-1 min-w-[160px]">
-                  <button
-                    type="button"
-                    onClick={() => handleSnooze(SNOOZE_DURATION_1H_MS, '1h')}
-                    className="w-full px-3 py-1.5 text-left text-xs hover:bg-secondary transition-colors"
-                  >
-                    In 1 hour
-                  </button>
-                  <button
-                    type="button"
-                    onClick={() => handleSnooze(SNOOZE_DURATION_1D_MS, '1d')}
-                    className="w-full px-3 py-1.5 text-left text-xs hover:bg-secondary transition-colors"
-                  >
-                    Tomorrow
-                  </button>
-                  <button
-                    type="button"
-                    onClick={() => handleSnooze(SNOOZE_DURATION_1W_MS, '1w')}
-                    className="w-full px-3 py-1.5 text-left text-xs hover:bg-secondary transition-colors"
-                  >
-                    Next week
-                  </button>
-                </div>
-              )}
-            </div>
-          </div>
-
-          <div className="flex items-center gap-2">
-            <button
-              type="button"
-              onClick={onClose}
-              className="px-4 py-1.5 text-sm text-muted-foreground hover:text-foreground transition-colors"
-            >
-              Later
-            </button>
-            <button
-              type="button"
-              onClick={handleUpdate}
-              disabled={updating}
-              className="px-4 py-1.5 text-sm font-medium bg-green-500 text-white rounded-lg hover:bg-green-600 disabled:opacity-50 transition-colors flex items-center gap-2"
-            >
-              {updating ? (
-                <>{t('status.updatingEllipsis')}</>
-              ) : (
-                <>
-                  <Download className="w-4 h-4" />
-                  {t('updates.updateNow')}
-                </>
-              )}
-            </button>
-          </div>
-        </div>
+        <WhatsNewFooter
+          updating={updating}
+          onSnooze={handleSnooze}
+          onSkip={handleSkip}
+          onUpdate={handleUpdate}
+          onClose={onClose}
+        />
       </BaseModal.Footer>
     </BaseModal>
   )
